feat(ui): migrate WPF to Avalonia for cross-platform support
