+++ AvaloniaUI/ViewModels/ViewModelBase.cs
@@ -1,0 +1,8 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace AvaloniaUI.ViewModels
+{
+    public abstract class ViewModelBase : ObservableObject
+    {
+    }
+}
