--- WPF-UI1/App.xaml
@@ -1,56 +1,0 @@
-﻿<Application x:Class="WPF_UI1.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-             xmlns:converters="clr-namespace:WPF_UI1.Converters"
-             xmlns:sys="clr-namespace:System;assembly=mscorlib"
-             StartupUri="MainWindow.xaml">
-    
-    <Application.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <!-- Material Design -->
-                <materialDesign:BundledTheme BaseTheme="Dark" PrimaryColor="Blue" SecondaryColor="Orange" />
-                <ResourceDictionary Source="pack://application:,,,/MaterialDesignThemes.Wpf;component/Themes/MaterialDesignTheme.Defaults.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-
-            <!-- 值转换器 -->
-            <converters:BooleanToColorConverter x:Key="BooleanToColorConverter"/>
-            <converters:BooleanToYesNoConverter x:Key="BooleanToYesNoConverter"/>
-            <converters:BooleanToSupportConverter x:Key="BooleanToSupportConverter"/>
-            <converters:BytesToStringConverter x:Key="BytesToStringConverter"/>
-            <converters:NetworkSpeedConverter x:Key="NetworkSpeedConverter"/>
-            <converters:DiskUsagePercentageConverter x:Key="DiskUsagePercentageConverter"/>
-
-            <!-- 深色主题样式 -->
-            <Style TargetType="Window">
-                <Setter Property="Background" Value="#FF353535"/>
-                <Setter Property="Foreground" Value="White"/>
-            </Style>
-
-            <Style TargetType="TextBlock">
-                <Setter Property="Foreground" Value="White"/>
-            </Style>
-
-            <Style TargetType="GroupBox">
-                <Setter Property="Foreground" Value="White"/>
-                <Setter Property="BorderBrush" Value="#FF555555"/>
-            </Style>
-
-            <!-- 自定义样式 -->
-            <Style x:Key="InfoLabelStyle" TargetType="TextBlock">
-                <Setter Property="Foreground" Value="#FFB3B3B3"/>
-                <Setter Property="FontSize" Value="12"/>
-                <Setter Property="VerticalAlignment" Value="Center"/>
-            </Style>
-
-            <Style x:Key="ValueLabelStyle" TargetType="TextBlock">
-                <Setter Property="Foreground" Value="White"/>
-                <Setter Property="FontSize" Value="12"/>
-                <Setter Property="FontWeight" Value="Normal"/>
-                <Setter Property="VerticalAlignment" Value="Center"/>
-            </Style>
-
-        </ResourceDictionary>
-    </Application.Resources>
-</Application>
